Move sharedcomponent to its own package
The goal is to remove from coreinternal everything that depends on the core (and not only on model), to
ensure that packages like the translators do not depend on anything more than just model.
