+++ internal/sharedcomponent/sharedcomponent.go
